+++ CHANGELOG.md
@@ -8,6 +8,11 @@ See [keep a changelog] for information about writing changes to this log.
 
 ## [Unreleased]
 
+* [PR-53](https://github.com/itk-dev/giv-din-stemme/pull/53)
+  Added commands for qualifying donations via whisper.
+* [PR-52](https://github.com/itk-dev/giv-din-stemme/pull/52)
+  * Added admin UI for `GivDinStemme` entities
+  * Added `validated` property to `GivDinStemme` entity
 * [PR-51](https://github.com/itk-dev/giv-din-stemme/pull/51)
   Removed email field.
 * [PR-50](https://github.com/itk-dev/giv-din-stemme/pull/50)